+++ orm/table_create_test.go
@@ -22,27 +22,27 @@ type CreateTableModel struct {
 	Uint64         uint64
 	Float32        float32
 	Float64        float64
-	Decimal        float64 `sql:"type:'decimal(10,10)'"`
+	Decimal        float64 `pg:"type:'decimal(10,10)'"`
 	ByteSlice      []byte
 	ByteArray      [32]byte
-	String         string    `sql:"default:'D\\'Angelo'"`
-	Varchar        string    `sql:",type:varchar(500)"`
-	Time           time.Time `sql:"default:now()"`
+	String         string    `pg:"default:'D\\'Angelo'"`
+	Varchar        string    `pg:",type:varchar(500)"`
+	Time           time.Time `pg:"default:now()"`
 	Duration       time.Duration
-	NotNull        int `sql:",notnull"`
+	NotNull        int `pg:",notnull"`
 	NullBool       sql.NullBool
 	NullFloat64    sql.NullFloat64
 	NullInt64      sql.NullInt64
 	NullString     sql.NullString
 	Slice          []int
-	SliceArray     []int `sql:",array"`
+	SliceArray     []int `pg:",array"`
 	Map            map[int]int
-	MapHstore      map[int]int `sql:",hstore"`
+	MapHstore      map[int]int `pg:",hstore"`
 	Struct         struct{}
 	StructPtr      *struct{}
-	Unique         int `sql:",unique"`
-	UniqueField1   int `sql:"unique:field1_field2"`
-	UniqueField2   int `sql:"unique:field1_field2"`
+	Unique         int `pg:",unique"`
+	UniqueField1   int `pg:"unique:field1_field2"`
+	UniqueField2   int `pg:"unique:field1_field2"`
 	JSONRawMessage json.RawMessage
 }
 
@@ -52,43 +52,43 @@ type CreateTableWithoutPKModel struct {
 
 type CreateTableOnDeleteOnUpdateModel struct {
 	ID                 int
-	CreateTableModelID int `sql:"on_delete:RESTRICT, on_update:CASCADE"`
+	CreateTableModelID int `pg:"on_delete:RESTRICT, on_update:CASCADE"`
 	CreateTableModel   *CreateTableModel
 }
 
 type CreateTableWithTablespace struct {
-	tableName string `sql:"tablespace:ssd"`
+	tableName string `pg:"tablespace:ssd"`
 
 	String string
 }
 
 type CreateTableWithRangePartition struct {
-	tableName string `sql:"partitionBy:RANGE (time)"`
+	tableName string `pg:"partitionBy:RANGE (time)"`
 
 	Time   time.Time
 	String string
 }
 
 type CreateTableWithListPartition struct {
-	tableName string `sql:"partitionBy:LIST (country)"`
+	tableName string `pg:"partitionBy:LIST (country)"`
 
 	Country string
 	String  string
 }
 
 type CreateTableWithHashPartition struct {
-	tableName string `sql:"partitionBy:HASH (account_id)"`
+	tableName string `pg:"partitionBy:HASH (account_id)"`
 
-	ID        int `sql:",pk,type:int,default:0"`
+	ID        int `pg:",pk,type:int,default:0"`
 	AccountID int
 	String    string
 }
 
 type CreateTableWithMultipleNamedUniques struct {
-	ID               int    `sql:",pk"`
-	AccountID        int    `sql:",unique:'per_account,per_store'"`
-	OrderNumber      string `sql:",unique:per_account"`
-	StoreOrderNumber string `sql:",unique:per_store"`
+	ID               int    `pg:",pk"`
+	AccountID        int    `pg:",unique:'per_account,per_store'"`
+	OrderNumber      string `pg:",unique:per_account"`
+	StoreOrderNumber string `pg:",unique:per_store"`
 }
 
 var _ = Describe("CreateTable", func() {